ODT - parsing headings
