+++ cr3qt/src/cr3widget.cpp
@@ -1247,13 +1247,13 @@ void CR3View::OnLoadFileFormatDetected( doc_format_t fileFormat )
         case doc_format_epub:
             filename = "epub.css";
             break;
-        case doc_format_odt:
         case doc_format_html:
             filename = "htm.css";
             break;
         case doc_format_doc:
             filename = "doc.css";
             break;
+        case doc_format_odt:
         case doc_format_docx:
             filename = "docx.css";
             break;
